feat: Move "Add Party" button to bottom of section
The "Add Party" button for the "Strane u Ugovoru" section has been moved from its original position at the top to the bottom of the section. This change improves the user interface by placing the primary action for adding parties at the end of the list, which is more intuitive for users adding multiple entries.